discord-bot/requirements.txt
# Discord Bot Core
discord.py==2.3.2
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# AI Integration
anthropic==0.8.1

# Cache
redis==5.0.1
aioredis==2.0.1

# HTTP Client
aiohttp==3.9.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.0
isort==5.13.2
pylint==3.0.3
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
